requirements.txt
numpy
matplotlib
tensorflow.keras.datasets